Extended types with a subtype of 0 are allowed. Change encoding to fit this fact.
git-svn-id: http://svn.mkgmap.org.uk/mkgmap/trunk@2139 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/typ/ShapeStacking.java
@@ -36,7 +36,7 @@ public class ShapeStacking {
 		int levelType = (level << 16) + type;
 		DrawOrder order = bar.get(levelType);
 		if (order == null) {
-			order = new DrawOrder((byte) type);
+			order = new DrawOrder(type);
 			bar.put(levelType, order);
 		}
